Improve workspace and cleanup functions * Remove some old commands which are no longer needed with ZFS.

+++ build.sh
@@ -16,7 +16,6 @@ fi
 cache="${livecd}/${arch}/cache"
 base="${cache}/${version}/base"
 packages="${cache}/packages"
-ports="${cache}/furybsd-ports-master"
 iso="${livecd}/iso"
   if [ -n "$CIRRUS_CI" ] ; then
     # On Cirrus CI ${livecd} is in tmpfs for speed reasons
@@ -71,20 +70,15 @@ isopath="${iso}/${vol}-${arch}.iso"
 
 workspace()
 {
-  umount ${uzip}/var/cache/pkg >/dev/null 2>/dev/null || true
-  umount ${ports} >/dev/null 2>/dev/null || true
-  rm -rf ${ports} >/dev/null 2>/dev/null || true
-  umount ${cache}/furybsd-packages/ >/dev/null 2>/dev/null || true
-  rm ${cache}/master.zip >/dev/null 2>/dev/null || true
-  umount ${uzip}/dev >/dev/null 2>/dev/null || true
-  zpool destroy furybsd >/dev/null 2>/dev/null || true
-  mdconfig -d -u 0 >/dev/null 2>/dev/null || true
-  if [ -f "${livecd}/pool.img" ] ; then
-    rm ${livecd}/pool.img
-  fi
-  if [ -d "${livecd}" ] ;then
-    # chflags -R noschg ${uzip} ${cdroot} >/dev/null 2>/dev/null || true
-    rm -rf ${uzip} ${cdroot} ${ports} >/dev/null 2>/dev/null || true
+  if [ -n "$CIRRUS_CI" ] ; then
+  # On CI systems there is no reason to clean up which takes time
+    return 0
+  else
+    umount ${uzip}/var/cache/pkg >/dev/null 2>/dev/null || true
+    umount ${uzip}/dev >/dev/null 2>/dev/null || true
+    zpool destroy furybsd >/dev/null 2>/dev/null || true
+    mdconfig -d -u 0 >/dev/null 2>/dev/null || true
+    rm ${livecd}/pool.img >/dev/null 2>/dev/null || true
   fi
   mkdir -p "${livecd}" "${base}" "${iso}" "${packages}" "${uzip}" "${ramdisk_root}/dev" "${ramdisk_root}/etc" >/dev/null 2>/dev/null
   truncate -s 3g "${livecd}/pool.img"
@@ -231,7 +225,7 @@ ramdisk()
   touch "${ramdisk_root}/etc/fstab"
   cp ${uzip}/etc/login.conf ${ramdisk_root}/etc/login.conf
   makefs -b '10%' "${cdroot}/data/ramdisk.ufs" "${ramdisk_root}"
-  gzip "${cdroot}/data/ramdisk.ufs"
+  gzip -f "${cdroot}/data/ramdisk.ufs"
   rm -rf "${ramdisk_root}"
 }
 
@@ -245,20 +239,18 @@ image()
 {
   sh "${cwd}/scripts/mkisoimages-${arch}.sh" -b "${label}" "${isopath}" "${cdroot}"
   md5 "${isopath}" > "${isopath}.md5"
+  echo "$isopath created"
 }
 
 cleanup()
 {
-if [ ! -z "${CI}" ] ; then
-  zpool destroy -f furybsd
-  # On CI systems there is no reason to clean up which takes time
-  return
-fi
-  if [ -d "${livecd}" ] ; then
-    # chflags -R noschg ${uzip} ${cdroot} >/dev/null 2>/dev/null
-    rm -rf ${uzip} ${cdroot} ${ports} >/dev/null 2>/dev/null
+  if [ -n "$CIRRUS_CI" ] ; then
+    # On CI systems there is no reason to clean up which takes time
+    return 0
+  else
+    zpool destroy -f furybsd
+    mdconfig -d -u 0
   fi
-  echo "$isopath created"
 }
 
 workspace
